Adding tests for IsDateMoreRecent() in HelperMethodsTests
