+++ Kinvey.Tests/ClientTests/ClientIntegrationTests.cs
