+++ Kinvey.Tests/ClientTests/ClientUnitTests.cs
