+++ Kinvey.Tests/DataStoreTests/DataStoreAutoIntegrationTests.cs
