+++ Kinvey.Tests/FileTests/FileIntegrationTests.cs
