+++ Kinvey.Tests/FileTests/FileUnitTests.cs
