+++ Kinvey.Tests/UserTests/UserUnitTests.cs
